Moving structs for config to a model package
