+++ internal/config/config.go
@@ -1,16 +1,12 @@
 package config
 
 import (
+	"github.com/emmaLP/gs-software-onboarding/internal/model"
 	"github.com/pkg/errors"
 	"github.com/spf13/viper"
 )
 
-type Configuration struct {
-	Consumer ConsumerConfig `mapstructure:",squash"`
-	Database DatabaseConfig `mapstructure:",squash"`
-}
-
-func LoadConfig(path string) (*Configuration, error) {
+func LoadConfig(path string) (*model.Configuration, error) {
 	v := viper.New()
 	v.SetConfigFile(path + "/.env")
 	v.AutomaticEnv()
@@ -18,7 +14,7 @@ func LoadConfig(path string) (*Configuration, error) {
 	if err := v.ReadInConfig(); err != nil {
 		return nil, errors.Wrap(err, "Failed to read env file")
 	}
-	var configuration Configuration
+	var configuration model.Configuration
 	err := v.Unmarshal(&configuration)
 	if err != nil {
 		return nil, errors.Wrap(err, "Unable to decode into map")
