--- internal/config/consumer.go
@@ -1,5 +1,0 @@
-package config
-
-type ConsumerConfig struct {
-	BaseUrl string `mapstructure:"base_url"`
-}
